elements of a2 appear in a1 consecutively
Take a number as input to two arrays a1 and a2 of sizes n and m respectively, where n >
m. Print “Yes” if a2 is contained in a1, else print 0. “Yes” should be printed if and only if
all the elements of a2 appear in a1 consecutively and in the same order, in any other
case, “No” should be printed.
Sample input: n = 7, m = 3
Array1
3 6 8 12 10 7 9 9 15
Array2
10 7 9
Sample output: yes